\name{dji30retw}
\docType{data}
\alias{dji30retw}
\title{data: Dow Jones 30 Constituents Closing Value log Weekly Return}
\description{
Dow Jones 30 Constituents closing value weekly (Friday) log returns from 
1987-03-16 to 2009-02-03 from Yahoo Finance. Note that AIG was replaced by KFT 
(Kraft Foods) on September 22, 2008. This is not reflected in this data set as 
that would bring the starting date of the data to 2001. When data was not 
available for a Friday, the closest previous close for which data was available 
was used.
}
\usage{data(dji30retw)}
\format{A data.frame containing 30x1141 observations.}
\source{Yahoo Finance}
\keyword{datasets}
