\name{mGARCHfilter-class}
\docType{class}
\alias{mGARCHfilter-class}

\title{Class: Multivariate GARCH Filter Class}
\description{High-level multivariate GARCH filter class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{\link[rugarch:GARCHfilter-class]{GARCHfilter}}, directly.
Class \code{\link[rugarch:rGARCH-class]{rGARCH}}, by class "GARCHfilter", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHfilter" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
