\name{mGARCHfit-class}
\docType{class}
\alias{mGARCHfit-class}
\title{Class: Multivariate GARCH Fit Class}
\description{High-level multivariate GARCH fit class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{GARCHfit} object from the \pkg{rugarch} package, directly.\\
Class \code{rGARCH} object from the \pkg{rugarch} package, by class \code{GARCHfit}, distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHfit" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
