\name{mGARCHforecast-class}
\docType{class}
\alias{mGARCHforecast-class}
\title{Class: Multivariate GARCH Forecast Class}
\description{High Level multivariate GARCH forecast class.}
\section{Objects from the Class}{A virtual Class: No objects may be created from 
it.}
\section{Extends}{
Class \code{"\linkS4class{GARCHforecast}"}, directly.
Class \code{"\linkS4class{rGARCH}"}, by class "GARCHforecast", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHforecast" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
