\name{mGARCHroll-class}
\docType{class}
\alias{mGARCHroll-class}
\title{Class: Multivariate GARCH Roll Class}
\description{High Level multivariate GARCH roll class.}
\section{Objects from the Class}{A virtual Class: No objects may be created from 
it.}
\section{Extends}{
Class \code{"\linkS4class{GARCHroll}"}, directly.
Class \code{"\linkS4class{rGARCH}"}, by class "GARCHroll", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHroll" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
