\name{mGARCHsim-class}
\docType{class}
\alias{mGARCHsim-class}
\title{Class: Multivariate GARCH Simulation Class}
\description{High-level multivariate GARCH simulation class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{GARCHsim} object from the \pkg{rugarch} package, directly.\\
Class \code{rGARCH} object from the \pkg{rugarch} package, by class \code{GARCHsim}, distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHsim" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
