\name{mGARCHspec-class}
\docType{class}
\alias{mGARCHspec-class}
\title{Class: Multivariate GARCH Specification}
\description{High Level multivariate GARCH specification class.}
\section{Objects from the Class}{A virtual Class: No objects may be created from 
it.}
\section{Extends}{
Class \code{"\linkS4class{GARCHspec}"}, directly.
Class \code{"\linkS4class{rGARCH}"}, by class "GARCHspec", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHspec" in the signature.
}
\author{Alexios Ghalanos}
\keyword{classes}
