\name{mGARCHfilter-class}
\Rdversion{1.1}
\docType{class}
\alias{mGARCHfilter-class}

\title{Class: Multivariate GARCH Filter Class}
\description{High Level multivariate GARCH filter class.}
\section{Objects from the Class}{A virtual Class: No objects may be created from 
it.}
\section{Extends}{
Class \code{"\linkS4class{GARCHfilter}"}, directly.
Class \code{"\linkS4class{rGARCH}"}, by class "GARCHfilter", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHfilter" in the signature.
}
\author{Alexios Ghalanos}
\keyword{classes}
