\name{mGARCHfilter-class}
\docType{class}
\alias{mGARCHfilter-class}

\title{Class: Multivariate GARCH Filter Class}
\description{High-level multivariate GARCH filter class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{GARCHfilter} object from the \pkg{rugarch} package, directly.\\
Class \code{rGARCH} object from the \pkg{rugarch} package, by class \code{GARCHfilter}, distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHfilter" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
