\name{mGARCHfit-class}
\docType{class}
\alias{mGARCHfit-class}
\title{Class: Multivariate GARCH Fit Class}
\description{High-level multivariate GARCH fit class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{\link[rugarch:GARCHfit-class]{GARCHfit}}, directly.
Class \code{\link[rugarch:rGARCH-class]{rGARCH}}, by class "GARCHfit", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHfit" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
