\name{mGARCHfit-class}
\docType{class}
\alias{mGARCHfit-class}
\title{Class: Multivariate GARCH Fit Class}
\description{High Level multivariate GARCH fit class.}
\section{Objects from the Class}{A virtual Class: No objects may be created from 
it.}
\section{Extends}{
Class \code{"\linkS4class{GARCHfit}"}, directly.
Class \code{"\linkS4class{rGARCH}"}, by class "GARCHfit", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHfit" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
