\name{mGARCHforecast-class}
\docType{class}
\alias{mGARCHforecast-class}
\title{Class: Multivariate GARCH Forecast Class}
\description{High-level multivariate GARCH forecast class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{\link[rugarch:GARCHforecast-class]{GARCHforecast}}, directly.
Class \code{\link[rugarch:rGARCH-class]{rGARCH}}, by class "GARCHforecast", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHforecast" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
