\name{mGARCHforecast-class}
\docType{class}
\alias{mGARCHforecast-class}
\title{Class: Multivariate GARCH Forecast Class}
\description{High-level multivariate GARCH forecast class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{GARCHforecast} object from the \pkg{rugarch} package, directly.\\
Class \code{rGARCH} object from the \pkg{rugarch} package, by class \code{GARCHforecast}, distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHforecast" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
