\name{mGARCHroll-class}
\docType{class}
\alias{mGARCHroll-class}
\title{Class: Multivariate GARCH Roll Class}
\description{High-level multivariate GARCH roll class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{GARCHroll} object from the \pkg{rugarch} package, directly.\\
Class \code{rGARCH} object from the \pkg{rugarch} package, by class \code{GARCHroll}, distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHroll" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
