\name{mGARCHsim-class}
\docType{class}
\alias{mGARCHsim-class}
\title{Class: Multivariate GARCH Simulation Class}
\description{High-level multivariate GARCH simulation class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{\link[rugarch:GARCHsim-class]{GARCHsim}}, directly.
Class \code{\link[rugarch:rGARCH-class]{rGARCH}}, by class "GARCHsim", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHsim" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
