\name{mGARCHsim-class}
\docType{class}
\alias{mGARCHsim-class}
\title{Class: Multivariate GARCH Simulation Class}
\description{High Level multivariate GARCH simulation class.}
\section{Objects from the Class}{A virtual Class: No objects may be created from 
it.}
\section{Extends}{
Class \code{"\linkS4class{GARCHsim}"}, directly.
Class \code{"\linkS4class{rGARCH}"}, by class "GARCHsim", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHsim" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
