\name{mGARCHspec-class}
\docType{class}
\alias{mGARCHspec-class}
\title{Class: Multivariate GARCH Specification}
\description{High-level multivariate GARCH specification class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{\link[rugarch:GARCHspec-class]{GARCHspec}}, directly.
Class \code{\link[rugarch:rGARCH-class]{rGARCH}}, by class "GARCHspec", distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHspec" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
