\name{mGARCHspec-class}
\docType{class}
\alias{mGARCHspec-class}
\title{Class: Multivariate GARCH Specification}
\description{High-level multivariate GARCH specification class.}
\section{Objects from the Class}{A virtual class: no objects may be created from it.}
\section{Extends}{
Class \code{GARCHspec} object from the \pkg{rugarch} package, directly.\\
Class \code{rGARCH} object from the \pkg{rugarch} package, by class \code{GARCHspec}, distance 2.
}
\section{Methods}{
No methods defined with class "mGARCHspec" in the signature.
}
\author{Alexios Galanos}
\keyword{classes}
